+++ src/Responses/PreauthCompleteResponse.php
@@ -15,6 +15,46 @@ class PreauthCompleteResponse extends BaseResponse
     }
 
     /**
+     * Get bank transaction ID
+     */
+    public function getBankTransaction(): ?string
+    {
+        return $this->data['bank_transaction'] ?? null;
+    }
+
+    /**
+     * Get bank response
+     */
+    public function getBankResponse(): ?string
+    {
+        return $this->data['bank_response'] ?? null;
+    }
+
+    /**
+     * Get RRN (Retrieval Reference Number)
+     */
+    public function getRrn(): ?string
+    {
+        return $this->data['rrn'] ?? null;
+    }
+
+    /**
+     * Get card mask
+     */
+    public function getCardMask(): ?string
+    {
+        return $this->data['card_mask'] ?? null;
+    }
+
+    /**
+     * Get cardholder name
+     */
+    public function getCardName(): ?string
+    {
+        return $this->data['card_name'] ?? null;
+    }
+
+    /**
      * Get amount
      */
     public function getAmount(): ?float