+++ src/Responses/RefundResponse.php
@@ -53,4 +53,12 @@ class RefundResponse extends BaseResponse
     {
         return isset($this->data['amount']) ? (float) $this->data['amount'] : null;
     }
+
+    /**
+     * Get bank response
+     */
+    public function getBankResponse(): ?string
+    {
+        return $this->data['bank_response'] ?? null;
+    }
 }